refactor: Use typescript in ApplicationHeader
