+++ src/domain/header/ApplicationHeader.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { ReactComponent as AhtiLogo } from '../../assets/icons/ahti_logo.svg';
-import DropdownMenu from '../DropdownMenu/DropdownMenu';
-import UnstyledLink from '../UnstyledLink/UnstyledLink';
+import DropdownMenu from '../../Components/DropdownMenu/DropdownMenu';
+import UnstyledLink from '../../Components/UnstyledLink/UnstyledLink';
 import { useTranslation } from 'react-i18next';
 
 import styled from 'styled-components';
@@ -39,8 +39,6 @@ const ApplicationHeader = () => {
           document.body.style.overflowY = 'scroll';
           setMenuOpen(false);
         }}
-        onLanguageFI={() => i18n.changeLanguage('fi')}
-        onLanguageEN={() => i18n.changeLanguage('en')}
       />
     </nav>
   );
